+++ lib/capybara/spec/session/fill_in_spec.rb
@@ -180,4 +180,24 @@ Capybara::SpecHelper.spec "#fill_in" do
       end.to raise_error(Capybara::ElementNotFound)
     end
   end
+
+  context "with { :clear => :backspace } fill_option", :requires => [:js] do
+    it 'should only trigger onchange once' do
+      @session.visit('/with_js')
+      @session.fill_in('with_change_event', :with => 'some value',
+                       :fill_options => { :clear => :backspace })
+      # click outside the field to trigger the change event
+      @session.find(:css, 'body').click
+      expect(@session.find(:css, '.change_event_triggered', :match => :one)).to have_text 'some value'
+    end
+
+    it 'should trigger change when clearing field' do
+      @session.visit('/with_js')
+      @session.fill_in('with_change_event', :with => '',
+                       :fill_options => { :clear => :backspace })
+      # click outside the field to trigger the change event
+      @session.find(:css, 'body').click
+      expect(@session).to have_selector(:css, '.change_event_triggered', :match => :one)
+    end
+  end
 end